+++ comment-service/src/main/java/com/gradproject/commentservice/repository/CommentRepositoryCustom.java
@@ -6,5 +6,5 @@ import java.util.List;
 
 public interface CommentRepositoryCustom {
 
-    List<CommentDto> getCommentDtoListBy(Long gameId);
+    List<CommentDto> getCommentDtoListBy(Integer gameId);
 }
